update SC BE & add render folder
