+++ maro/simulator/scenarios/supply_chain/business_engine.py
@@ -1,7 +1,8 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
-import collections
+
 import os
+from collections import defaultdict
 from typing import Dict, List
 
 from maro.backends.frame import FrameBase
@@ -155,8 +156,8 @@ class SupplyChainBusinessEngine(AbsBusinessEngine):
         actions = event.payload
         assert isinstance(actions, list)
 
-        consumer_actions_by_unit: Dict[int, List[ConsumerAction]] = collections.defaultdict(list)
-        manufacture_actions_by_unit: Dict[int, List[ManufactureAction]] = collections.defaultdict(list)
+        consumer_actions_by_unit: Dict[int, List[ConsumerAction]] = defaultdict(list)
+        manufacture_actions_by_unit: Dict[int, List[ManufactureAction]] = defaultdict(list)
         for action in actions:
             if isinstance(action, ConsumerAction):
                 consumer_actions_by_unit[action.id].append(action)
@@ -202,8 +203,8 @@ class SupplyChainBusinessEngine(AbsBusinessEngine):
                     facility.id: {
                         "in_transit_orders": facility.get_in_transit_orders(),
                         "pending_order":
-                            None if facility.distribution is None
-                            else facility.distribution.get_pending_product_quantities(),
+                            defaultdict(int) if facility.distribution is None
+                            else facility.distribution.pending_product_quantity,
                     } for facility in self.world.facilities.values()
                 }
             }
